rename process_adapters
